stresize

+++ ft_stresize.c
@@ -6,13 +6,13 @@
 /*   By: rreedy <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/18 11:38:53 by rreedy            #+#    #+#             */
-/*   Updated: 2018/08/18 11:39:26 by rreedy           ###   ########.fr       */
+/*   Updated: 2018/08/18 11:45:37 by rreedy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <libft.h>
+#include "libft.h"
 
-char	*ft_stresize(char **s, int in, int size)
+char	*ft_stresize(char **s, int in, size_t size)
 {
 	char	*str;
